Add files via upload
Let's party problem


Your friend visits you and says he has an invitation to a party
at which there will be celebrities.  He is very excited by the
prospect of meeting famous people.  However, he has a problem.
These celebrities are going to attend the party at different time
periods. For example two individuals may attend the party for three
hours but these may be at non overlapping intervals.
Even though he knows when each celebrity will be attending,
your friend is not able to figure out when to attend the party such
that he meets the most number of celebrities.  The problem is also that he
has only 59 minutes to attend the party given the approaching finals week.

You say Aha!  I can write a program that gets a solution for you.
Tell me when each celebrity is going to be at the party and my
program will tell you what's the best time to
visit the party.

Assume a 24 hour clock and at least 5 celebrities

Assume that the durations always start on the hour and end just
before an hour. For example, a celebrity may enter the party at 11
and leave just before 14 hours. This means the person has been there
for just short of three hours.  Individuals cannot enter at other times
or leave at other times.  This holds for your friend as well.
Assume too that each celebrity attends the party only once.

Write a program that tells your friend which hour is the best in which to
join the party given each celebrity's entry and exit hours.

Two basic questions as always.

How to represent the problem? How to solve the problem?